# Conflicts:
#	src/main/java/com/solegendary/reignofnether/ability/EnchantEquipAbilityServerboundPacket.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/EnchantMaiming.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/EnchantMultishot.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/EnchantQuickCharge.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/EnchantSharpness.java
#	src/main/java/com/solegendary/reignofnether/ability/abilities/EnchantVigor.java
#	src/main/java/com/solegendary/reignofnether/building/Building.java
#	src/main/java/com/solegendary/reignofnether/building/BuildingClientEvents.java
#	src/main/java/com/solegendary/reignofnether/building/BuildingServerEvents.java
#	src/main/java/com/solegendary/reignofnether/building/RangeIndicator.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/monsters/Laboratory.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/Bastion.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/piglins/InfernalPortal.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/placements/HealingFountainPlacement.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/placements/LibraryPlacement.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/placements/RangeIndicatorProductionPlacement.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/placements/SculkCatalystPlacement.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/placements/StrongholdPlacement.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/Castle.java
#	src/main/java/com/solegendary/reignofnether/building/buildings/villagers/Library.java
#	src/main/java/com/solegendary/reignofnether/entities/BlazeUnitFireball.java
#	src/main/java/com/solegendary/reignofnether/minimap/MinimapClientEvents.java
#	src/main/java/com/solegendary/reignofnether/orthoview/OrthoviewClientEvents.java
#	src/main/java/com/solegendary/reignofnether/time/TimeClientEvents.java
#	src/main/java/com/solegendary/reignofnether/unit/UnitActionItem.java
#	src/main/java/com/solegendary/reignofnether/unit/UnitClientEvents.java
#	src/main/java/com/solegendary/reignofnether/unit/units/villagers/WitchUnit.java

+++ src/main/java/com/solegendary/reignofnether/unit/goals/UnitBowAttackGoal.java
@@ -51,12 +51,12 @@ public class UnitBowAttackGoal<T extends net.minecraft.world.entity.Mob> extends
             this.attackCooldown -= 1;
     }
 
-    public int getAttackCooldown() {
+    public float getAttackCooldown() {
         return attackCooldown;
     }
 
     public void setToMaxAttackCooldown() {
-        this.attackCooldown = ((AttackerUnit) this.mob).getAttackCooldown();
+        this.attackCooldown = (int) ((AttackerUnit) this.mob).getAttackCooldown();
     }
 
     public void resetCooldown() {
