+++ src/main/java/com/solegendary/reignofnether/unit/goals/UnitRangedAttackGoal.java
@@ -45,12 +45,12 @@ public class UnitRangedAttackGoal<T extends net.minecraft.world.entity.Mob> exte
             this.attackCooldown -= 1;
     }
 
-    public int getAttackCooldown() {
+    public float getAttackCooldown() {
         return attackCooldown;
     }
 
     public void setToMaxAttackCooldown() {
-        this.attackCooldown = ((AttackerUnit) this.mob).getAttackCooldown();
+        this.attackCooldown = (int) ((AttackerUnit) this.mob).getAttackCooldown();
     }
 
     public void resetCooldown() {
